Squashed 'ipython3-versioncontrol/' changes from 22a7e89..b71e7b5
b71e7b5 Move log messages to jupyter logger. Jupyter basic operations on notebooks are done on its tornado logger, using stdout instead can cause the saving process to fail if stdout channel was closed (in the case of an ssh connection for instance). In that case the script would crash on the print() function.
REVERT: 22a7e89 Move log messages to stderr. Jupyter basic operations on notebooks are done on stderr and using stdout instead can cause the saving process to fail because the channel was closed (in the case of an ssh connection for example) and the script crashed on the print() function.

git-subtree-dir: ipython3-versioncontrol
git-subtree-split: b71e7b5e6f2de6087d3147852310a2fee5c50ece